build: latest semantic-release
I reviewed the release notes and the only thing I am worried about is
the `branch` to `branches` change. As far as I can tell this repo and
`telus-semantic-release-config` to not use it. The plugins that our
shared config do use may need to be updated but I am unsure.